parsing: remove potential panics
Removes the `unwrap` calls in `BigEndian*::from_bytes` implementations.

Currently, the implementation will never panic on the `unwrap` call,
because the prior `get` call always returns the appropriate length byte
slice.

However, it is safer to call `.ok()?` to bubble-up the `None` result.
Which, again, will never be returned with the current implementation.
The function will bubble-up `None` on the `get` call before reaching the
`try_into`.